+++ api/project.js
@@ -8,7 +8,9 @@ const {
   modifyProject,
   deleteProject,
   getProjectTeamMembers,
+  getAuthorUniqueId,
 } = require('../dao/projectDAO');
+const jwtVerify = require('../utils/jwtVerify');
 
 const router = express.Router();
 
@@ -41,35 +43,45 @@ router.get('/:project_id', async (req, res) => {
   }
 });
 
-router.post('/', async (req, res) => {
-  try {
-    const result = await createProject(req.body);
-    res.status(201).json(result);
-  } catch (error) {
-    res.status(500).json({ responseMessage: 'failure', responseData: null, error });
-  }
+router.post('/', (req, res) => {
+  jwtVerify(req, res, async ({ user_id: currentUserId }) => {
+    if (currentUserId === req.body.userUserId) {
+      const result = await createProject(req.body);
+      res.status(201).json(result);
+    } else {
+      res.status(401).json({ responseMessage: 'Warning! Use only your own tokens' });
+    }
+  });
 });
 
-router.patch('/:project_id', async (req, res) => {
-  const { project_id } = req.params;
+router.patch('/:project_id', (req, res) => {
+  jwtVerify(req, res, async decoded => {
+    const { user_id: currentUserId } = decoded;
+    const { project_id } = req.params;
 
-  try {
-    const result = await modifyProject({ ...req.body, projectId: project_id });
-    res.status(200).json({ responseMessage: 'success', responseData: result });
-  } catch (error) {
-    res.send({ responseMessage: 'failure', responseData: null, error });
-  }
+    const [{ user_user_id: authorId }] = await getAuthorUniqueId(project_id);
+
+    if (currentUserId === authorId) {
+      const result = await modifyProject({ ...req.body, projectId: project_id, currentUserId });
+      res.status(200).json({ responseMessage: 'success', responseData: result });
+    } else {
+      res.status(401).json({ responseMessage: 'Warning! Use only your own tokens' });
+    }
+  });
 });
 
-router.delete('/:project_id', async (req, res) => {
-  const { project_id } = req.params;
-
-  try {
-    const result = await deleteProject(project_id);
-    res.status(200).json({ responseMessage: 'success', responseData: result });
-  } catch (error) {
-    res.send({ responseMessage: 'failure', responseData: null, error });
-  }
+router.delete('/:project_id', (req, res) => {
+  jwtVerify(req, res, async decoded => {
+    const { user_id: currentUserId } = decoded;
+    const { project_id } = req.params;
+    const [{ user_user_id: authorId }] = await getAuthorUniqueId(project_id);
+    if (currentUserId === authorId) {
+      const result = await deleteProject(project_id);
+      res.status(200).json({ responseMessage: 'success', responseData: result });
+    } else {
+      res.status(401).json({ responseMessage: 'Warning! Use only your own tokens' });
+    }
+  });
 });
 
 module.exports = router;
