+++ api/projectLike.js
@@ -1,6 +1,7 @@
 const express = require('express');
 const { postLike, deleteLike, checkExistedLike } = require('../dao/likeDAO');
 const { getProjectLikeCount } = require('../dao/projectDAO');
+const jwtVerify = require('../utils/jwtVerify');
 
 const router = express.Router();
 
@@ -13,39 +14,43 @@ router.get('/', async (req, res) => {
   }
 });
 
-router.post('/', async (req, res) => {
-  try {
-    const { affectedRows } = await postLike(req.query);
-    if (affectedRows) {
-      try {
-        const [likeCount] = await getProjectLikeCount(req.query.project_id);
-        res.status(200).json({ responseMessage: 'success', likeCount: likeCount.likeCount });
-      } catch (error) {
-        res.status(500).json({ responseMessage: 'failure', error });
-      }
+router.post('/', (req, res) => {
+  jwtVerify(req, res, async ({ user_id: currentUserId }) => {
+    if (+req.query.user_id === currentUserId) {
+      const { affectedRows } = await postLike(req.query);
+      if (affectedRows) {
+        try {
+          const [likeCount] = await getProjectLikeCount(req.query.project_id);
+          res.status(200).json({ responseMessage: 'success', likeCount: likeCount.likeCount });
+          return;
+        } catch (error) {
+          res.status(500).json({ responseMessage: 'failure', error });
+        }
+      } else res.status(400).json({ responseMessage: 'failure' });
+    } else {
+      res.status(401).json({ responseMessage: 'Warning! Use only your own tokens' });
     }
-  } catch (error) {
-    res.status(500).json({ responseMessage: 'failure', error });
-  }
-  res.status(400).json({ responseMessage: 'failure' });
+  });
 });
 
-router.delete('/', async (req, res) => {
-  try {
-    const { affectedRows } = await deleteLike(req.query);
+router.delete('/', (req, res) => {
+  jwtVerify(req, res, async ({ user_id: currentUserId }) => {
+    if (+req.query.user_id === currentUserId) {
+      const { affectedRows } = await deleteLike(req.query);
 
-    if (affectedRows) {
-      try {
-        const [likeCount] = await getProjectLikeCount(req.query.project_id);
-        res.status(200).json({ responseMessage: 'success', likeCount: likeCount.likeCount });
-      } catch (error) {
-        res.status(500).json({ responseMessage: 'failure', error });
-      }
+      if (affectedRows) {
+        try {
+          const [likeCount] = await getProjectLikeCount(req.query.project_id);
+          res.status(200).json({ responseMessage: 'success', likeCount: likeCount.likeCount });
+          return;
+        } catch (error) {
+          res.status(500).json({ responseMessage: 'failure', error });
+        }
+      } else res.status(400).json({ responseMessage: 'The users likes do not exist for this post.' });
+    } else {
+      res.status(401).json({ responseMessage: 'Warning! Use only your own tokens' });
     }
-  } catch (error) {
-    res.status(500).json({ responseMessage: 'failure', error });
-  }
-  res.status(400).json({ responseMessage: 'The users likes do not exist for this post.' });
+  });
 });
 
 module.exports = router;
